dump openstack::test_file timeout to 60 seconds
I previously saw a lot of failures in the
openstack::test_file on red hat systems.

I have bumped this timeout up (the time that we
wait between assigned a floating ip and trying to
route to it). This seems to have made this error
go away most of the time.