Add ability to configure installed openstack version
perform a hiera lookup for the openstack_version
to use to setup the appropriate packages(for ubuntu).

Ensure that setup/hosts.pp is run before (and not from
within) setup/os.pp. This is b/c hiera is not
required to be installed before the os specific
setup manifest is run.