This code refactors the openstack test script
This script was updated so that I could test swift as well
as cherry-pick commits from openstack's gerrit git repo.

it adds conditionally logic for when the repo is swift (to
build out the swift test env which includes a puppetmaster)

it also adds a test_mode=puppet_swift which runs the basic
swift smoke tests.

it also adds code so that I can run the gerrit cherry-pick command
to test patches from gerrit that are being reviewed.

Change-Id: If944160dfcc6b68b56f4f688d9d6dea314bf0c5a