fix gui_mode
Previously gui_mode was checking for a value
of 'true' in the config.yaml file.

This did not match the default value of true
that was set there.

This commit changes this parameter to expect
a boolean value of true or false instead of a
string to make it more consistent with the other
values.

Change-Id: If3f55a16e38bd4bea310bac31b9ef7ac9de158af