update rabbit_password
Previously, we were not setting the rabbit_user of compute.

This led to test failures when we changed the default
of rabbit_user in the openstack module.

This commit adds rabbit_user to the list of things configured
in the hiera file and ensures that the compute nodes set their
rabbit_user correctly.

Change-Id: Ifb9b17f2e96dd66cea2566d7353f14b2d727fa83