update rabbit_password
Previously, we were not setting the rabbit_user of
compute instances.

This led to test failures when we changed the default
of rabbit_user in the openstack module.

This commit adds rabbit_user to the list of things configured
in the hiera file and ensures that the compute nodes set their
rabbit_user correctly.

Change-Id: Ic89c5f856a00d6bab2c3cab4b8a74818afc29fb4